+++ R/utils.R
@@ -49,3 +49,14 @@ map_chr <- function(.x, .f, ...) {
 map_int <- function(.x, .f, ...) {
   vapply(.x, .f, FUN.VALUE = integer(1), ...)
 }
+
+any_na <- function(x) {
+  t <- typeof(x)
+  if (t == "list") {
+    .Call(nanoparquet_any_null, x)
+  } else if (t == "double") {
+    .Call(nanoparquet_any_na, x)
+  } else {
+    anyNA(x)
+  }
+}